+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.0.0] - 2021-08-30
+
+### Changed
+- Removed deprecated "DecryptData" methods.
+- Some refactoring in SplitKeyEncryption.
+
 ## [5.5.1] - 2021-08-13
 
 ### Changed